Empty reply when rate limit grpc service uses certain command operators
If certain command operators are used as the value of an initial_metadata
header to be sent to with gRPC requests to rate limit service then Envoy
fails to process a request and returns an empty body.

Correct behaviour without initial_metadata:

   $ git checkout main
   $ curl -H "foo: a" -H "bar: b" http://0.0.0.0:8888/twoheader
     Hello World

Broken behaviour with initial_metadata (and certain command operators);

   $ git checkout rate_limit_service-route_name-empty-reply-example
   $ docker-compose -f docker-compose-example.yml up
   $ curl -H "foo: a" -H "bar: b" http://0.0.0.0:8888/twoheader
     curl: (52) Empty reply from server

This happens if you use certain command operators in the value of
a key defined in `initial_metadata` e.g.

    rate_limit_service:
      grpc_service:
        envoy_grpc:
          cluster_name: ratelimit
        initial_metadata:
          - key: my-lovely-key
            value: "%ROUTE_NAME%"

The following command operators also exhibit the same behaviour:

- %UPSTREAM_CLUSTER%

The following command operators *do not* exhibit the same behaviour:

- %REQ(:PATH)%